Export features and device extensions from the device module instead of the instance module
