+++ examples/src/bin/msaa-renderpass.rs
@@ -86,7 +86,6 @@ use vulkano::framebuffer::Subpass;
 use vulkano::image::AttachmentImage;
 use vulkano::image::Dimensions;
 use vulkano::image::StorageImage;
-use vulkano::instance::Features;
 use vulkano::instance::Instance;
 use vulkano::instance::InstanceExtensions;
 use vulkano::instance::PhysicalDevice;
@@ -103,7 +102,7 @@ fn main() {
         .find(|&q| q.supports_graphics())
         .expect("couldn't find a graphical queue family");
     let (device, mut queues) = {
-        Device::new(physical, &Features::none(), &DeviceExtensions::none(),
+        Device::new(physical, physical.supported_features(), &DeviceExtensions::none(),
                     [(queue_family, 0.5)].iter().cloned()).expect("failed to create device")
     };
     let queue = queues.next().unwrap();
